fastapi-backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
motor==3.1.1
pymongo==4.3.3
pydantic==1.10.7
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
